Call google.charts.Line.convertOptions(options)
https://developers.google.com/chart/interactive/docs/gallery/linechart?csw=1#creating-material-line-charts says:

The Material Charts are in beta. The appearance and interactivity are
largely final, but the way options are declared is not. If you are
converting a Classic Line Chart to a Material Line Chart, you'll want to
replace this line:

chart.draw(data, options);

...with this:

chart.draw(data, google.charts.Line.convertOptions(options));